split into sorted array stats
