+++ libs/nmbrs_statistics/src/descriptive_stats/mod.rs
@@ -1,1 +1,6 @@
 mod array_stats;
+mod sorted_array_stats;
+
+pub use array_stats::{
+    covariance, mean, sum, variance, Covariance, Mean, Sum, Variance, VarianceBias,
+};
